# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample server EHRS server.",
    "version": "1.0.0",
    "title": "EHRS",
    "contact": {
      "email": "Admin@ehrs.ir"
    }
  },
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/api/products/getAll": {
      "post": {
        "tags": [
          "Product"
        ],
        "summary": "Return all product by filters",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Product filters",
            "schema": {
              "type": "object",
              "required": [
                "size",
                "page"
              ],
              "properties": {
                "size": {
                  "type": "integer",
                  "example": 10
                },
                "page": {
                  "type": "integer",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/api/products/create": {
      "post": {
        "tags": [
          "Product"
        ],
        "summary": "Create product item",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Product object that needs to insert",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response"
          },
          "401": {
            "description": "product validation failed"
          }
        }
      }
    },
...
    "/api/products/update": {
      "post": {
        "tags": [
          "Product"
        ],
        "summary": "update product item by _id",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Product object that needs to update",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response"
          },
          "401": {
            "description": "product validation failed"
          }
        }
      }
    },
...
    "/api/products/import": {
      "post": {
        "tags": [
          "Product"
        ],
        "summary": "import product items from xlsx file",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "products",
            "in": "formData",
            "description": "Updated file of the product",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response"
          },
          "401": {
            "description": "product validation failed"
          }
        }
      }
    },
...
    "/api/products/export": {
      "post": {
        "tags": [
          "Product"
        ],
        "summary": "export product items xlsx file",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Product filters that needs to export",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response"
          },
          "401": {
            "description": "product validation failed"
          }
        }
      }
    },
    "/api/products/delete": {
      "post": {
        "tags": [
          "Product"
        ],
        "summary": "delete product item by _id",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "_id of the product that needs to be deleted",
            "schema": {
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "5f180141e7d39f2f5f53bd1d"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response"
          },
          "401": {
            "description": "product validation failed"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
    "Product": {
      "type": "object",
      "required": [
        "eRx",
        "packageCode"
      ],
      "properties": {
        "eRx": {
          "type": "string",
          "example": "eRx123466"
        },
        "packageCode": {
          "type": "string",
          "example": "A"
        },
        "eRxGeneric": {
          "type": "string",
          "example": "eRx464545"
        },
        "sepasCode": {
          "type": "string",
          "example": "2000000"
        },
        "genericCode": {
          "type": "integer",
          "example": 45
        },
        "packageCount": {
          "type": "integer",
          "example": 100
        },
        "enGenericName": {
          "type": "string",
          "example": "Sertraline"
        },
        "faGenericName": {
          "type": "string",
          "example": "سرترالین"
        },
        "enBrandName": {
          "type": "string",
          "example": "Zoloft"
        },
        "faBrandName": {
          "type": "string",
          "example": "زولوفت"
        },
        "licenceOwner": {
          "type": "string",
          "example": "داروپخش"
        },
        "brandOwner": {
          "type": "string",
          "example": "داروپخش"
        },
        "producer": {
          "type": "string",
          "example": "داروپخش"
        },
        "countryBrandOwner": {
          "type": "string",
          "example": "ایران"
        },
        "countryProducer": {
          "type": "string",
          "example": "ایران"
        },
        "gtn": {
          "type": "array",
          "example": [
            "123456998855"
          ]
        },
        "irc": {
          "type": "array",
          "example": [
            "123456998855"
          ]
        },
        "category": {
          "type": "array",
          "example": [
            "A01B02"
          ]
        },
        "nativeIrc": {
          "type": "string",
          "example": "123456998855005"
        },
        "cName": {
          "type": "array",
          "example": [
            "ضدافسردگی"
          ]
        },
        "videos": {
          "type": "array",
          "example": [
            " https://www.aparat.com/v/WOS6C "
          ]
        },
        "properties": {
          "type": "array",
          "example": [
            {
              "key": "size",
              "value": "2.8nm"
            }
          ]
        }
      }
    },
...
  }
}